requirements.txt
allure-behave==2.9.45
allure-python-commons==2.9.45
Appium-Python-Client==1.0.2
attrs==21.4.0
behave==1.2.6
certifi==2020.12.5
chardet==3.0.4
docopt==0.6.2
et-xmlfile==1.1.0
idna==2.10
openpyxl==3.0.9
parse==1.18.0
parse-type==0.5.2
pipreqs==0.4.11
pluggy==1.0.0
PyYAML==6.0
requests==2.25.0
selenium==3.141.0
six==1.15.0
urllib3==1.26.2
yarg==0.1.9
